chore: restructure folders for CoursesPage components
